First working version of the heterogeneous birth-death process

+++ src/core/datatypes/phylogenetics/ratemap/RateMap_Biogeography.cpp
@@ -210,7 +210,7 @@ RateMap_Biogeography* RateMap_Biogeography::clone(void) const
 double RateMap_Biogeography::getRate(const TopologyNode& node, std::vector<CharacterEvent*> from, CharacterEvent* to, unsigned* count, double age) const
 {
     double rate = 0.0;
-    int s = to->getState();
+    size_t s = to->getState();
     
     if (from[ to->getCharacterIndex() ]->getState() == to->getState())
     {
@@ -271,7 +271,7 @@ double RateMap_Biogeography::getRate(const TopologyNode& node, std::vector<Chara
 double RateMap_Biogeography::getSiteRate(const TopologyNode& node, CharacterEvent* from, CharacterEvent* to, double age) const
 {
     double rate = 0.0;
-    int s = to->getState();
+    size_t s = to->getState();
 //    int charIdx = to->getIndex();
 //    int epochIdx = getEpochIndex(age);
     
@@ -297,7 +297,7 @@ double RateMap_Biogeography::getSiteRate(const TopologyNode& node, CharacterEven
 double RateMap_Biogeography::getSiteRate(const TopologyNode& node, size_t from, size_t to, size_t charIdx, double age) const
 {
     double rate = 0.0;
-    int s = to;
+    size_t s = to;
     
 //    int pres = !s;
 //    int epochIdx = getEpochIndex(age);
@@ -429,7 +429,7 @@ double RateMap_Biogeography::getUnnormalizedSumOfRates(const TopologyNode& node,
     double sum = 0.0;
     for (size_t i = 0; i < from.size(); i++)
     {
-        unsigned s = from[i]->getState();
+        size_t s = from[i]->getState();
         double v = availableAreaVector[ epochIdx * this->numCharacters + i ];
         
         if (forbidExtinction && s == 1 && counts[1] == 0)
